feat: add Orders page and update Navbar for order navigation
